+++ WindowsBase/EngineImplementation/XnaExtensions.cs
@@ -28,7 +28,7 @@ namespace MonoGameImplementation.EngineImplementation
 
         public static Rectangle ToXnaRect(this GameFramework.Rectangle rectangle)
         {
-            return new Rectangle(rectangle.X, rectangle.Y, rectangle.Width, rectangle.Height);
+            return new Rectangle((int)rectangle.X, (int)rectangle.Y, (int)rectangle.Width, (int)rectangle.Height);
         }
     }
 }
